updating milestone words count on successfull deletion

+++ back-end/controller/wordController/deleteWordController.js
@@ -1,6 +1,8 @@
+import Milestone from "../../models/Milestone.js";
 import Word from "../../models/Word.js";
 
 const deleteWord = async (req, res) => {
+  let milestone;
   try {
     // Extract user ID, word Id and milestone ID
     const userId = req.user.id;
@@ -13,6 +15,20 @@ const deleteWord = async (req, res) => {
         .status(400)
         .json({ message: "Milestone ID and Word ID are required." });
     }
+
+    // Find the milestone
+    milestone = await Milestone.findOne({ _id: milestoneId });
+    if (!milestone) {
+      return res
+        .status(404)
+        .json({ message: "No associated milestone was found!" });
+    }
+
+    // Update word count in milestone
+    await Milestone.updateOne(
+      { _id: milestoneId },
+      { $inc: { wordsCount: -1 } }
+    );
 
     // Find and delete the word
     const word = await Word.findOneAndDelete({
@@ -32,6 +48,13 @@ const deleteWord = async (req, res) => {
     res.status(200).json({ message: "Word deleted successfully." });
   } catch (error) {
     console.error("Error deleting word:", error);
+    // Rollback: Decrement wordsCount if the word save failed
+    if (milestone) {
+      await Milestone.updateOne(
+        { _id: milestone._id },
+        { $inc: { wordsCount: 1 } }
+      );
+    }
     res
       .status(500)
       .json({ message: "An error occurred while deleting the word." });
